requirements.txt
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.1
idna==2.10
progressbar2==4.0.0
python-utils==3.3.3
requests==2.23.0
urllib3==1.25.11
vk==2.0.2
